Split papi.d.ts
