+++ extensions/.eslintrc.cjs
@@ -1,9 +1,16 @@
 module.exports = {
   extends: '../.eslintrc.js',
   rules: {
-    strict: 'off',
     'import/no-unresolved': ['error', { ignore: ['papi'] }],
   },
+  overrides: [
+    {
+      files: ['*.js'],
+      rules: {
+        strict: 'off',
+      },
+    },
+  ],
   settings: {
     'import/resolver': {
       // See https://github.com/benmosher/eslint-plugin-import/issues/1396#issuecomment-575727774 for line below
