+++ extensions/lib/hello-world/hello-world.web-view.tsx
@@ -1,4 +1,4 @@
-import papi from 'papi';
+import papi from 'papi-frontend';
 import {
   Button,
   Checkbox,
