+++ lib/papi-dts/tsconfig.json
@@ -8,8 +8,8 @@
     "typeRoots": ["../node_modules/@types", "../extensions/dist", "../src/@types"]
   },
   "files": [
-    "../../src/renderer/services/papi-renderer.service.ts",
-    "../../src/extension-host/services/papi-extension-host.service.ts",
+    "../../src/renderer/services/papi-frontend.service.ts",
+    "../../src/extension-host/services/papi-backend.service.ts",
     "../../src/extension-host/extension-types/extension.interface.ts"
   ]
 }
